PullRequest: 368 [lite] Refactor train engine after merging contributions from GitHub
Merge branch fw/lite-train-engine of git@code.alipay.com:inclusionAI/AReaL.git into lite
https://code.alipay.com/inclusionAI/AReaL/pull_requests/368



* .
* .
