+++ arealite/tests/test_train_engine.py
@@ -13,7 +13,6 @@ from transformers import AutoTokenizer
 
 from arealite.api.cli_args import MicroBatchSpec, OptimizerConfig, TrainEngineConfig
 from arealite.api.io_struct import FinetuneSpec, SaveLoadMeta
-from arealite.engine.fsdp_engine import FSDPEngine
 
 VOCAB_SIZE = 100
 MODEL_PATH = "/storage/testing/models/Qwen__Qwen3-1.7B/"
@@ -53,10 +52,10 @@ def mock_input(
 
 
 def get_engine(engine_type: str, model_path: str):
+    from arealite.engine.autotp_engine import DeepSpeedAutoTPEngine
     from arealite.engine.fsdp_engine import FSDPEngine
-    from arealite.engine.hf_engine import HFEngine
 
-    engine_cls = {"hf": HFEngine, "fsdp": FSDPEngine}[engine_type]
+    engine_cls = {"auto_tp": DeepSpeedAutoTPEngine, "fsdp": FSDPEngine}[engine_type]
 
     engine_config = TrainEngineConfig(
         experiment_name=f"test-{engine_type}-engine",
@@ -75,7 +74,7 @@ def mock_loss_fn(logits: torch.Tensor, input_data: Dict) -> torch.Tensor:
     return torch.mean(logits)
 
 
-@pytest.fixture(scope="module", params=["fsdp", "hf"])
+@pytest.fixture(scope="module", params=["fsdp", "auto_tp"])
 def engine(request):
     os.environ.update(
         {
@@ -136,6 +135,12 @@ def test_train_batch(engine, mock_input):
 
 @torch.no_grad()
 def test_hf_save_load_weights(tmp_path_factory, engine, mock_input):
+    from arealite.engine.autotp_engine import DeepSpeedAutoTPEngine
+
+    if isinstance(engine, DeepSpeedAutoTPEngine):
+        print("AutoTP engine does not support HF save/load for now.")
+        return
+
     tokenizer = AutoTokenizer.from_pretrained(MODEL_PATH)
     path = tmp_path_factory.mktemp("hf_engine_test")
     save_load_meta = SaveLoadMeta(
